backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.7.2
bcrypt==4.1.3
python-dotenv==1.0.1
pillow==10.3.0
werkzeug==3.0.3
